# earth-frontend/openapi/getplanet-api.yaml
{"openapi":"3.1.0","info":{"title":"Planet Earth API","description":"OpenAPI Ingestion & Local Vector-Search Service","version":"0.1.0"},"paths":{"/v1/schemas":{"post":{"tags":["schemas"],"summary":"Upload an OpenAPI schema","description":"Upload a YAML or JSON OpenAPI 3.x specification file for processing.","operationId":"upload_schema_v1_schemas_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_schema_v1_schemas_post"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["schemas"],"summary":"List all schemas","description":"Retrieve a list of all uploaded schemas with their metadata.","operationId":"list_schemas_v1_schemas_get","parameters":[{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SchemaRead"},"title":"Response List Schemas V1 Schemas Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/schemas/{schema_id}":{"get":{"tags":["schemas"],"summary":"Get schema details","description":"Retrieve details about a specific schema by its ID.","operationId":"get_schema_v1_schemas__schema_id__get","parameters":[{"name":"schema_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Schema Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchemaRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["schemas"],"summary":"Delete schema","description":"Delete a schema and all its associated endpoints.","operationId":"delete_schema_v1_schemas__schema_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"schema_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Schema Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/search":{"post":{"tags":["search"],"summary":"Semantic search","description":"Search for API endpoints using natural language queries.","operationId":"search_endpoints_v1_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchQuery"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/EndpointSearchResult"},"type":"array","title":"Response Search Endpoints V1 Search Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/management/schemas/{schema_id}/endpoints":{"get":{"tags":["management"],"summary":"List endpoints for schema","description":"Retrieve all endpoints associated with a specific schema.","operationId":"list_schema_endpoints_v1_management_schemas__schema_id__endpoints_get","parameters":[{"name":"schema_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Schema Id"}},{"name":"include_deleted","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Deleted"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EndpointRead"},"title":"Response List Schema Endpoints V1 Management Schemas  Schema Id  Endpoints Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/management/reindex":{"post":{"tags":["management"],"summary":"Reindex vector store","description":"Reindex all endpoints in the vector store. This is an asynchronous operation.","operationId":"reindex_vector_store_v1_management_reindex_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"schema_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Schema Id"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/management/endpoints/{endpoint_id}/reindex":{"post":{"tags":["management"],"summary":"Reindex endpoint","description":"Reindex a specific endpoint in the vector store. This is an asynchronous operation.","operationId":"reindex_endpoint_api_v1_management_endpoints__endpoint_id__reindex_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"endpoint_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Endpoint Id"}}],"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_schema_v1_schemas_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_schema_v1_schemas_post"},"EndpointRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"schema_id":{"type":"string","format":"uuid","title":"Schema Id"},"path":{"type":"string","title":"Path"},"method":{"type":"string","title":"Method"},"operation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Operation Id"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"tags":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tags"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["schema_id","path","method"],"title":"EndpointRead","description":"Endpoint read model (pydantic)."},"EndpointSearchResult":{"properties":{"id":{"type":"string","title":"Id"},"schema_id":{"type":"string","title":"Schema Id"},"path":{"type":"string","title":"Path"},"method":{"type":"string","title":"Method"},"operation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Operation Id"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"tags":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tags"},"score":{"type":"number","title":"Score"},"schema_title":{"type":"string","title":"Schema Title"},"schema_version":{"type":"string","title":"Schema Version"},"spec":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Spec"}},"type":"object","required":["id","schema_id","path","method","score","schema_title","schema_version"],"title":"EndpointSearchResult","description":"Model for endpoint search results including vector similarity score."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SchemaRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"title":{"type":"string","title":"Title"},"version":{"type":"string","title":"Version"},"checksum":{"type":"string","title":"Checksum"},"status":{"type":"string","title":"Status"}},"type":"object","required":["title","version","checksum","status"],"title":"SchemaRead","description":"Schema read model (pydantic)."},"SearchQuery":{"properties":{"q":{"type":"string","title":"Q","description":"Search query text"},"top_k":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"Top K","description":"Number of results to return","default":5},"filter_schema_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filter Schema Id","description":"Optional filter by schema ID"},"filter_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filter Method","description":"Optional filter by HTTP method"},"include_deprecated":{"type":"boolean","title":"Include Deprecated","description":"Whether to include deprecated endpoints","default":false}},"type":"object","required":["q"],"title":"SearchQuery","description":"Model for semantic search requests."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}